pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "otel-distro-builder"
version = "1.0.0"
description = "Build and package custom OpenTelemetry Collector distributions"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.9"
authors = [
    { name = "ObservIQ" }
]
keywords = ["opentelemetry", "collector", "build", "manifest"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "google-cloud-storage==2.17.0",
    "pyyaml>=6.0.1",
    "colorlog>=6.8.0",
    "psutil>=6.1.0",
    "requests==2.32.2",
    "tqdm==4.66.3",
    "packaging==23.2",
    "cryptography>=42.0.0",
    "urllib3>=2.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black",
    "isort",
    "pylint",
    "mypy",
]
build = [
    "pyinstaller>=6.0",
]

[project.scripts]
otel-distro-builder = "builder.src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["builder*"]

[tool.setuptools.package-data]
builder = ["versions.yaml", "templates/*"]
"builder.src" = ["components.yaml", "bindplane_components.yaml"]
